requirements.txt
Flask>=3.0.0
Flask-WTF>=1.2.0
Flask-Caching>=2.1.0
Flask-SQLAlchemy>=3.1.0
Flask-Limiter>=3.5.0
Werkzeug>=3.0.0
psutil>=5.9.0
pyautogui>=0.9.54
pyperclip>=1.8.2
Pillow>=10.0.0
screeninfo>=0.8.1
mss>=9.0.0
requests>=2.31.0
PyYAML>=6.0.0
pywin32>=306; sys_platform == 'win32'